+++ NEWS.md
@@ -1,5 +1,7 @@
 # shinyOAuth (development version)
 
+* Prevent 'mirai' warning spam about 'stats' maybe not being available in workers.
+
 # shinyOAuth 0.3.0
 
 * Async backend: the default async backend is now 'mirai' (>= 2.0.0) for